Move add_entries() from moz.l10n.transform to moz.l10n.resource
