+++ moz/l10n/resource/__init__.py
@@ -1,3 +1,4 @@
+from .add_entries import add_entries
 from .format import Format, bilingual_extensions, detect_format, l10n_extensions
 from .l10n_equal import l10n_equal
 from .parse_resource import UnsupportedResource, parse_resource
@@ -6,6 +7,7 @@ from .serialize_resource import serialize_resource
 __all__ = [
     "Format",
     "UnsupportedResource",
+    "add_entries",
     "bilingual_extensions",
     "detect_format",
     "l10n_extensions",
