--- moz/l10n/transform/__init__.py
@@ -1,3 +1,0 @@
-from .add_entries import add_entries
-
-__all__ = ["add_entries"]
