+++ tests/test_add_entries.py
@@ -16,8 +16,8 @@ from __future__ import annotations
 
 from unittest import TestCase
 
+from moz.l10n.resource import add_entries
 from moz.l10n.resource.data import Entry, Resource, Section
-from moz.l10n.transform import add_entries
 
 
 class TestAddEntries(TestCase):
